refactor(client): move admin kv to separate page
