+++ src/routes/(manage)/admin/+page.server.ts
@@ -1,5 +1,3 @@
-import { BUSINESS_NAME } from "$env/static/private";
-import { createKey, getKeys } from "$lib/server/database/keys";
 import type { Actions } from "@sveltejs/kit";
 import type { PageServerLoad } from "./$types";
 import { generateMergedBilling } from "$lib/server/deal";
@@ -10,8 +8,6 @@ import {
 import { addMonths } from "date-fns";
 
 export const load: PageServerLoad = async ({ url }) => {
-	const keys = await getKeys(BUSINESS_NAME);
-
 	const yearFilter = url.searchParams.get("year");
 
 	const startDateFilter = yearFilter
@@ -31,26 +27,10 @@ export const load: PageServerLoad = async ({ url }) => {
 
 	const expectedWithSalesmen = await getExpectedWithSalesmen();
 
-	return { keys, payments, expected: expectedWithSalesmen };
+	return { payments, expected: expectedWithSalesmen };
 };
 
 export const actions = {
-	submit: async ({ request }) => {
-		const data = await request.formData();
-
-		const id = data.get("id") as string;
-		const key = data.get("key") as string;
-		const value = data.get("value") as string;
-
-		const newKey = await createKey({ business: BUSINESS_NAME, key, id, value });
-
-		return {
-			data: {
-				key: newKey.k,
-			},
-			method: id ? "update" : "insert",
-		};
-	},
 	printBilling: async () => {
 		const data = await generateMergedBilling("desc");
